% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_cone.R
\name{calculate_cone}
\alias{calculate_cone}
\title{Solid Geometry Calculator for cones}
\usage{
calculate_cone(radius_mm, height_mm)
}
\arguments{
\item{radius_mm}{The radius of the cone in millimeters.}

\item{height_mm}{The height of the cone in millimeters.}
}
\value{
Prints the diameter, scope, base area, line surface, lateral surface, surface and the volume of a cone for the inputs radius_mm und height_mm
}
\description{
Compute some key figures for a cone
}
\examples{
calculate_cone(radius_mm=50, height_mm=20)
}
