% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_cylinder.R
\name{calculate_cylinder}
\alias{calculate_cylinder}
\title{Solid Geometry Calculator for cylinders}
\usage{
calculate_cylinder(radius_mm, height_mm)
}
\arguments{
\item{radius_mm}{The radius of the cylinder in millimeters.}

\item{height_mm}{The height of the cylinder in millimeters.}
}
\value{
Prints the diameter, scope, base area, lateral surface, surface and the volume of a cylinder for the inputs radius_mm und height_mm
}
\description{
Compute some key figures for a cylinder
}
\examples{
calculate_cylinder(radius_mm=50, height_mm=20)

}
