% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_sphere.R
\name{calculate_sphere}
\alias{calculate_sphere}
\title{Solid Geometry Calculator for spheres}
\usage{
calculate_sphere(radius, unit = "mm")
}
\arguments{
\item{radius}{The radius of the sphere in millimeters. The User can select the input unit between millimeter ("mm"), centimeter ("cm") and meter ("m").}
}
\value{
Prints the diameter, scope, circular area, surface and the volume of a sphere for the input radius in the selected unit millimeter mm, centimeter cm or meter m.
}
\description{
Compute some key figures for a sphere
}
\examples{
calculate_sphere(radius=1.4, unit = "mm")
}
