% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_cone.R
\name{calculate_cone}
\alias{calculate_cone}
\title{Solid Geometry Calculator for cones}
\usage{
calculate_cone(radius, height, unit = "mm")
}
\arguments{
\item{radius}{The radius of the cone.}

\item{height}{The height of the cone.}

\item{unit}{The unit in which the key figures should be calculated. The User can select the input unit between millimeter ("mm"), centimeter ("cm") and meter ("m").}
}
\value{
Prints the diameter, scope, base area, line surface, lateral surface, surface and the volume of a cone for the inputs radius and height in the selected unit millimeter mm, centimeter cm or meter m.
}
\description{
Compute some key figures for a cone
}
\examples{
calculate_cone(radius=50, height=20, unit = "mm")
}
