% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_sphere.R
\name{calculate_sphere}
\alias{calculate_sphere}
\title{Solid Geometry Calculator for spheres}
\usage{
calculate_sphere(radius_mm)
}
\arguments{
\item{radius_mm}{The radius of the sphere in millimeters.}
}
\value{
Prints the diameter, scope, circular area, surface and the volume of a sphere for the input radius_mm
}
\description{
Compute some key figures for a sphere
}
\examples{
calculate_sphere(radius_mm=1.4)
}
